+++ lib/runners/processor.rb
@@ -225,6 +225,15 @@ module Runners
       end
     end
 
+    def add_warning_for_deprecated_linter(alternative:, deadline: nil)
+      deadline_str = deadline ? deadline.strftime("on %B %-d, %Y") : "in the near future"
+      add_warning <<~MSG.strip, file: ci_config_path_name
+        DEPRECATION WARNING!!!
+        The support for #{analyzer_name} is deprecated. Sider will drop these versions #{deadline_str}.
+        Please consider using an alternative tool #{alternative}.
+      MSG
+    end
+
     # Prohibit directory traversal attack, e.g.
     # - '../../etc/passwd'
     # - 'config/../../../etc/passwd'