+++ lib/runners/processor/gometalinter.rb
@@ -30,6 +30,7 @@ module Runners
         end
       end
 
+      add_warning_for_deprecated_linter(alternative: "GolangCi-Lint", deadline: Time.new(2_020, 3, 31))
       yield
     end
 
@@ -44,7 +45,6 @@ module Runners
           './...',
           '--deadline=1200s'
         )
-
         Results::Success.new(guid: guid, analyzer: analyzer).tap do |result|
           JSON.parse(stdout).each do |issue|
             loc = Location.new(